requirements.txt
setuptools>=65
wheel

Flask==2.3.3
Werkzeug==2.3.7
opencv-python-headless==4.8.1.78
numpy==1.26.4
ultralytics==8.0.200
google-generativeai==0.3.2
reportlab==4.0.4
python-dotenv==1.0.0
Pillow>=10.0.0
torch>=2.2.0
torchvision>=0.17.0

gunicorn==21.2.0
